backend/reservations/urls.py
# backend/reservations/urls.py
"""
URLs pour le système de réservation de médicaments (User Story 6)
"""
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from .views import ReservationViewSet

app_name = 'reservations'

router = DefaultRouter()
router.register(r'reservations', ReservationViewSet, basename='reservation')

urlpatterns = router.urls
